feat(quality): Night-17 타입 안전성 + Flutter 관측성/접근성/버그수정 (16건)
서버 (7건):
- stats.rs: i32 산술 전 f64 캐스팅, num_days as i32 → i32::try_from
- crawlers/mod.rs: f32 부동소수점 동시성 계산 → 정수 연산
- coupang.rs: LazyLock unwrap() → expect() 7곳 (명확한 패닉 메시지)
- auth_service.rs: jwt_refresh_ttl_secs as i64 → try_from 안전 변환 2곳
- auth_service.rs: referral code len() → chars().count() (유니코드 안전)
- reward_service.rs + pagination.rs: has_more usize 비교 방향 수정

Flutter (9건):
- product_detail_screen.dart: NumberFormat build()마다 생성 → static final
- product_detail_screen.dart: catch(e) → catch(e, st) + debugPrint
- home_screen.dart: catch(e) → catch(e, st) + 인기검색어 Semantics 접근성
- point_history_screen.dart: catch(e) → catch(e, st)
- point_history_screen.dart: _formatDate inline → static final DateFormat
- point_history_screen.dart: _transactionLabel referral_welcome_referrer 케이스 추가 (버그)
- my_page_screen.dart: _loadPoints/_doCheckin catch(e, st) 스택 트레이스
- alert_screen.dart: _loadAlerts catch(e) → catch(e, st) + debugPrint

검증: cargo test --lib 191/191 + clippy 0 warnings + flutter test 164/164 + analyze 0

+++ server/src/api/pagination.rs
@@ -43,7 +43,7 @@ impl<T: Serialize> PaginatedResponse<T> {
     /// `limit`개 초과 시 `has_more = true`로 설정하고 마지막 항목을 제거한다.
     /// `cursor_fn`은 마지막 가시 항목에서 커서 값을 추출하는 클로저.
     pub fn new(mut items: Vec<T>, limit: i64, cursor_fn: impl Fn(&T) -> String) -> Self {
-        let has_more = items.len() as i64 > limit;
+        let has_more = items.len() > limit as usize;
         if has_more {
             items.pop();
         }
